sistema/auth-service/requirements.txt
# auth-service/requirements.txt
Flask
requests
python-dotenv
gunicorn
PyJWT